refactor(google-meet): split plugin entry modules
